requirements.txt
fastapi==0.78.0
SQLAlchemy==1.4.23
uvicorn==0.15.0
asyncpg==0.24.0
GeoAlchemy2==0.9.4
Shapely==1.8.0
cachecall==0.1.0
httpx==0.23.0
pyshp==2.3.0
yoyo-migrations==7.3.2
psycopg2-binary==2.9.1
